backend/requirements.txt

langchain==0.3.26
fastapi==0.115.0
uvicorn==0.32.0
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone==0.2.8
langchain-google-genai==2.0.6
langchain-community==0.3.26
pydantic==2.10.2
pydantic-settings==2.6.1
python-multipart==0.0.12



